fix(build): break package dependency SCC
