+++ plugins/plugin-personal-assistant/src/automation-node-contributor.error-policy.test.ts
@@ -14,7 +14,7 @@ import { describe, expect, it, vi } from "vitest";
 
 // The module under test pulls a value import from app-core's registration seam;
 // stub it so this unit test does not require app-core to be built.
-vi.mock("@elizaos/app-core/api/automation-node-contributors", () => ({
+vi.mock("@elizaos/shared/automation-node-contributors", () => ({
   registerAutomationNodeContributor: vi.fn(),
 }));
